feat: map shortcut for diff commands, add license in code file.
fixed:
  - wrong spells in my name, add license GPLv3.0 in code file.
  - when match xml/html filetype, auto-pairs can work fine.
  - binding buffer/tab almost like one-by-one, map sp/vs shortcut.
  - modify some options to make vim operating comfortable.

About vim buffer and tabs:

This is not how Vim's tabs are designed to be used.
In fact, they're misnamed. A better name would be "viewport" or "layout",
because that's what a tab is-
it's a different layout of windows of all of your existing buffers.

Trying to beat Vim into 1 tab == 1 buffer is an exercise in futility.
Vim doesn't know or care and it will not respect it on all
commands-in particular, anything that uses the quickfix buffer.
(:make, :grep, and :helpgrep are the ones that spring to mind)
will happily ignore tabs and there's nothing you can do to stop that.